+++ Avatar/CBase64.h
@@ -16,22 +16,22 @@ using std::vector;
 class AVATAR_EXPORT CBase64 {
 public:
 	//! 编码 Base64 字符串
-	static int Encode(const unsigned char* data, int length, char* buffer, int size);
+	static size_t Encode(const unsigned char* data, size_t length, char* buffer, size_t size);
 	//! 编码 Base64 文本
-	static int Encode(const unsigned char* data, int length, string& buffer);
+	static size_t Encode(const unsigned char* data, size_t length, string& buffer);
 	//! 编码 Base64 字符串
-	static int Encode(const vector<unsigned char>& data, char* buffer, int size);
+	static size_t Encode(const vector<unsigned char>& data, char* buffer, size_t size);
 	//! 编码 Base64 文本
-	static int Encode(const vector<unsigned char>& data, string& buffer);
+	static size_t Encode(const vector<unsigned char>& data, string& buffer);
 
 	//! 解码 Base64 字符串
-	static int Decode(const char* data, unsigned char* buffer, int size);
+	static size_t Decode(const char* data, unsigned char* buffer, size_t size);
 	//! 解码 Base64 文本
-	static int Decode(const string& data, unsigned char* buffer, int size);
+	static size_t Decode(const string& data, unsigned char* buffer, size_t size);
 	//! 解码 Base64 字符串
-	static int Decode(const char* data, vector<unsigned char>& buffer);
+	static size_t Decode(const char* data, vector<unsigned char>& buffer);
 	//! 解码 Base64 文本
-	static int Decode(const string& data, vector<unsigned char>& buffer);
+	static size_t Decode(const string& data, vector<unsigned char>& buffer);
 };
 
 #endif
